requirements.txt
click
erddapy
geopandas
joblib
matplotlib
netCDF4
numpy
opencv-python
pandas
Pillow
psutil
pyrosm
rembg
requests
scikit-learn
seaborn
shapely
torch
torchvision
tqdm
xgboost
xarray
